fix: address unit test for splitting bedrock embedding invoke flow into helper methods
